@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -36,6 +36,7 @@ dependencies:
   firebase_core: ^1.14.0
   dialogflow_flutter: ^1.0.0
   firebase_storage: ^10.2.10
+  file_picker: ^4.5.1
 
 dev_dependencies:
   flutter_test:
